Apply a function to each element of a vector

The dt_map functions transform their input by applying a function to each element and returning a vector the same length as the input.

dt_map() returns a list

_lgl(), _int(), _dbl() and _chr() variants return their specified type

_dfr() & _dfc() Return data frame results binded together

Note: the '~' alias will not work with any of the dt_map() functions

dt_map2_dfc(.x, .y, .f, ...)

Arguments

.x
A list or vector

.f
A function

...
Other arguments to pass to a function

Value

Examples

1:10 %>% map(rnorm, n = 10)

#> Error in map(., rnorm, n = 10): could not find function "map"